New Update (#3)
* CariHareketler ve Fatura veritabanını birleştirildi. Gerekli değişiklikler yapıldı.

* CariHareketler ve Fatura veritabanını birleştirildi. Gerekli değişiklikler yapıldı.
Ufak iyileştirmeler yapıldı.

* feat: MySQL bağlantı ayar ekranı eklendi ve ini dosyası yönetimi geliştirildi

- Kullanıcıların MySQL sunucusu bilgilerini manuel olarak girebileceği `MySqlSettingsWindow` arayüzü eklendi.
- Bağlantı ayarları için `mysqlsettings.ini` dosyası AppData dizinine otomatik olarak oluşturuluyor.
- Veritabanı bağlantı bilgilerinin (server, port, user, password) ini dosyasından okunması sağlandı.
- Şifre güvenliği için password değeri Base64 ile şifrelenerek saklanıyor.
- `ConfigManager`, ini dosyasından dinamik veri alacak şekilde güncellendi.
- Veritabanı bağlantısı başarısız olursa ayar penceresi otomatik olarak açılıyor.
- Program başlarken bağlantı kontrolü yapılıyor, kullanıcı giriş yaptıktan sonra program kapanmadan devam ediyor.

* Ufak iyileştirmeler

+++ Cari kayıt Programı/PositiveToColorConverter.cs
@@ -9,7 +9,7 @@ namespace Cari_kayıt_Programı
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            if (value is double amount)
+            if (value is decimal amount)
             {
                 if (amount > 0)
                 {
